[LCR-221] refactor: 코드 분리
